+++ sources/code/Plugins/GraphicsVulkan/VulkanPipeline.cpp
@@ -153,7 +153,7 @@ namespace Grindstone {
 			colorBlending.sType = VK_STRUCTURE_TYPE_PIPELINE_COLOR_BLEND_STATE_CREATE_INFO;
 			colorBlending.logicOpEnable = VK_FALSE;
 			colorBlending.logicOp = VK_LOGIC_OP_COPY;
-			colorBlending.attachmentCount = colorBlendAttachments.size();
+			colorBlending.attachmentCount = static_cast<uint32_t>(colorBlendAttachments.size());
 			colorBlending.pAttachments = colorBlendAttachments.data();
 			colorBlending.blendConstants[0] = 0.0f;
 			colorBlending.blendConstants[1] = 0.0f;
